+++ usage-dcc-location-callback-nav/pages/produce.py
@@ -1,5 +1,0 @@
-import dash
-
-
-def layout(prodcode=None, **other_unknown_query_strings):
-    return dash.html.Div(f"Produce dashboard for product code {prodcode}")